+++ lex.yy.c
@@ -545,10 +545,11 @@ char *yytext;
     #include "../parser.tab.hpp"
     #include <string.h> 
     #include <stdlib.h> 
+    #include <string>
     int yylex();
-#line 550 "lex.yy.c"
+#line 551 "lex.yy.c"
 
-#line 552 "lex.yy.c"
+#line 553 "lex.yy.c"
 
 #define INITIAL 0
 #define COMMENT 1
@@ -766,9 +767,9 @@ YY_DECL
 		}
 
 	{
-#line 15 "app/lexer.l"
+#line 16 "app/lexer.l"
 
-#line 772 "lex.yy.c"
+#line 773 "lex.yy.c"
 
 	while ( /*CONSTCOND*/1 )		/* loops until end-of-file is reached */
 		{
@@ -837,207 +838,207 @@ do_action:	/* This label is used only to access EOF actions. */
 
 case 1:
 YY_RULE_SETUP
-#line 16 "app/lexer.l"
+#line 17 "app/lexer.l"
 {BEGIN(COMMENT);}
 	YY_BREAK
 case 2:
 YY_RULE_SETUP
-#line 17 "app/lexer.l"
+#line 18 "app/lexer.l"
 {BEGIN(INITIAL);}
 	YY_BREAK
 case 3:
 /* rule 3 can match eol */
 YY_RULE_SETUP
-#line 18 "app/lexer.l"
+#line 19 "app/lexer.l"
 ;
 	YY_BREAK
 case 4:
 YY_RULE_SETUP
-#line 19 "app/lexer.l"
+#line 20 "app/lexer.l"
 {return (VAR); }
 	YY_BREAK
 case 5:
 YY_RULE_SETUP
-#line 20 "app/lexer.l"
+#line 21 "app/lexer.l"
 { return (_BEGIN); }
 	YY_BREAK
 case 6:
 YY_RULE_SETUP
-#line 21 "app/lexer.l"
+#line 22 "app/lexer.l"
 { return (ENDDO); }
 	YY_BREAK
 case 7:
 YY_RULE_SETUP
-#line 22 "app/lexer.l"
+#line 23 "app/lexer.l"
 { return (ENDWHILE); }
 	YY_BREAK
 case 8:
 YY_RULE_SETUP
-#line 23 "app/lexer.l"
+#line 24 "app/lexer.l"
 { return (ENDFOR); }
 	YY_BREAK
 case 9:
 YY_RULE_SETUP
-#line 24 "app/lexer.l"
+#line 25 "app/lexer.l"
 { return (END); }
 	YY_BREAK
 case 10:
 YY_RULE_SETUP
-#line 25 "app/lexer.l"
+#line 26 "app/lexer.l"
 { return (REPEAT);}
 	YY_BREAK
 case 11:
 YY_RULE_SETUP
-#line 26 "app/lexer.l"
+#line 27 "app/lexer.l"
 { return (UNTIL);}
 	YY_BREAK
 case 12:
 YY_RULE_SETUP
-#line 27 "app/lexer.l"
+#line 28 "app/lexer.l"
 { return (ASSIGN); }
 	YY_BREAK
 case 13:
 YY_RULE_SETUP
-#line 28 "app/lexer.l"
+#line 29 "app/lexer.l"
 { return (IF); }
 	YY_BREAK
 case 14:
 YY_RULE_SETUP
-#line 29 "app/lexer.l"
+#line 30 "app/lexer.l"
 { return (THEN); }
 	YY_BREAK
 case 15:
 YY_RULE_SETUP
-#line 30 "app/lexer.l"
+#line 31 "app/lexer.l"
 { return (ELSE); }
 	YY_BREAK
 case 16:
 YY_RULE_SETUP
-#line 31 "app/lexer.l"
+#line 32 "app/lexer.l"
 { return (ENDIF); }
 	YY_BREAK
 case 17:
 YY_RULE_SETUP
-#line 32 "app/lexer.l"
+#line 33 "app/lexer.l"
 { return (WHILE); }
 	YY_BREAK
 case 18:
 YY_RULE_SETUP
-#line 33 "app/lexer.l"
+#line 34 "app/lexer.l"
 { return (DO); }
 	YY_BREAK
 case 19:
 YY_RULE_SETUP
-#line 34 "app/lexer.l"
+#line 35 "app/lexer.l"
 { return (FOR); }
 	YY_BREAK
 case 20:
 YY_RULE_SETUP
-#line 35 "app/lexer.l"
+#line 36 "app/lexer.l"
 { return (FROM); }
 	YY_BREAK
 case 21:
 YY_RULE_SETUP
-#line 36 "app/lexer.l"
+#line 37 "app/lexer.l"
 { return (READ); }
 	YY_BREAK
 case 22:
 YY_RULE_SETUP
-#line 37 "app/lexer.l"
+#line 38 "app/lexer.l"
 { return (WRITE); }
 	YY_BREAK
 case 23:
 YY_RULE_SETUP
-#line 38 "app/lexer.l"
+#line 39 "app/lexer.l"
 { return (PLUS); }
 	YY_BREAK
 case 24:
 YY_RULE_SETUP
-#line 39 "app/lexer.l"
+#line 40 "app/lexer.l"
 { return (MINUS); }
 	YY_BREAK
 case 25:
 YY_RULE_SETUP
-#line 40 "app/lexer.l"
+#line 41 "app/lexer.l"
 { return (TIMES); }
 	YY_BREAK
 case 26:
 YY_RULE_SETUP
-#line 41 "app/lexer.l"
+#line 42 "app/lexer.l"
 { return (DIV); }
 	YY_BREAK
 case 27:
 YY_RULE_SETUP
-#line 42 "app/lexer.l"
+#line 43 "app/lexer.l"
 { return (MOD); }
 	YY_BREAK
 case 28:
 YY_RULE_SETUP
-#line 43 "app/lexer.l"
+#line 44 "app/lexer.l"
 { return (EQ); }
 	YY_BREAK
 case 29:
 YY_RULE_SETUP
-#line 44 "app/lexer.l"
+#line 45 "app/lexer.l"
 { return (NEQ); }
 	YY_BREAK
 case 30:
 YY_RULE_SETUP
-#line 45 "app/lexer.l"
+#line 46 "app/lexer.l"
 { return (LE); }
 	YY_BREAK
 case 31:
 YY_RULE_SETUP
-#line 46 "app/lexer.l"
+#line 47 "app/lexer.l"
 { return (GE); }
 	YY_BREAK
 case 32:
 YY_RULE_SETUP
-#line 47 "app/lexer.l"
+#line 48 "app/lexer.l"
 { return (LEQ); }
 	YY_BREAK
 case 33:
 YY_RULE_SETUP
-#line 48 "app/lexer.l"
+#line 49 "app/lexer.l"
 { return (GEQ); }
 	YY_BREAK
 case 34:
 YY_RULE_SETUP
-#line 49 "app/lexer.l"
+#line 50 "app/lexer.l"
 {yylval.direction = false; return (direction); }
 	YY_BREAK
 case 35:
 YY_RULE_SETUP
-#line 50 "app/lexer.l"
+#line 51 "app/lexer.l"
 {yylval.direction = true; return (direction); }
 	YY_BREAK
 case 36:
 YY_RULE_SETUP
-#line 51 "app/lexer.l"
+#line 52 "app/lexer.l"
 {yylval.pidentifier = (char*) strdup(yytext); return (pidentifier);}
 	YY_BREAK
 case 37:
 YY_RULE_SETUP
-#line 52 "app/lexer.l"
-{yylval.num = atoi(yytext); return(num);}
+#line 53 "app/lexer.l"
+{yylval.num = std::stoll(yytext); return(num);}
 	YY_BREAK
 case 38:
 /* rule 38 can match eol */
 YY_RULE_SETUP
-#line 53 "app/lexer.l"
+#line 54 "app/lexer.l"
 ;
 	YY_BREAK
 case 39:
 YY_RULE_SETUP
-#line 54 "app/lexer.l"
+#line 55 "app/lexer.l"
 { return yytext[0]; }
 	YY_BREAK
 case 40:
 YY_RULE_SETUP
-#line 55 "app/lexer.l"
+#line 56 "app/lexer.l"
 ECHO;
 	YY_BREAK
-#line 1041 "lex.yy.c"
+#line 1042 "lex.yy.c"
 case YY_STATE_EOF(INITIAL):
 case YY_STATE_EOF(COMMENT):
 	yyterminate();
@@ -2055,5 +2056,5 @@ void yyfree (void * ptr )
 
 #define YYTABLES_NAME "yytables"
 
-#line 55 "app/lexer.l"
+#line 56 "app/lexer.l"
